Use BigDecimal instead of Double to store middle value of currency rate

+++ app/src/main/java/com/example/currencyradar/domain/models/CurrentRate.kt
@@ -1,6 +1,8 @@
 package com.example.currencyradar.domain.models
 
+import java.math.BigDecimal
+
 data class CurrentRate(
     val currency: Currency,
-    val middleValue: Double,
+    val middleValue: BigDecimal,
 )
